+++ plugins/gui/include/gui/graph_widget/graph_graphics_view.h
@@ -115,6 +115,7 @@ namespace hal
 
         void handleModuleDialog();
         void handleCancelPickModule();
+        void handleShortestPath();
 
     private:
         void paintEvent(QPaintEvent* event) override;